Finished implementing EVApproachingPartner pred
